@@ -5,25 +5,24 @@
   "main": "index.js",
   "license": "Apache-2.0",
   "devDependencies": {
-    "@babel/core": "^7.17.9",
-    "@babel/eslint-parser": "^7.17.0",
-    "@babel/plugin-proposal-class-properties": "^7.16.7",
-    "@babel/preset-env": "^7.16.11",
-    "@babel/preset-react": "^7.16.7",
-    "babel-loader": "^8.2.4",
-    "cypress": "^9.5.4",
+    "@babel/core": "^7.20.12",
+    "@babel/eslint-parser": "^7.19.1",
+    "@babel/plugin-proposal-class-properties": "^7.18.6",
+    "@babel/preset-env": "^7.20.2",
+    "@babel/preset-react": "^7.18.6",
+    "babel-loader": "^8.3.0",
+    "cypress": "^9.7.0",
     "eslint": "^7.32.0",
-    "eslint-config-prettier": "^8.5.0",
+    "eslint-config-prettier": "^8.6.0",
     "eslint-plugin-cypress": "^2.12.1",
     "eslint-plugin-ignore-generated-and-nolint": "^1.0.0",
-    "eslint-plugin-jsx-a11y": "^6.5.1",
-    "eslint-plugin-react": "^7.29.4",
-    "pixelmatch": "^5.2.1",
+    "eslint-plugin-jsx-a11y": "^6.7.1",
+    "eslint-plugin-react": "^7.32.1",
+    "pixelmatch": "^5.3.0",
     "pngjs": "^6.0.0",
-    "prettier": "^2.6.2",
-    "react-select": "^2.4.4",
-    "webpack": "^5.72.0",
-    "webpack-cli": "^4.9.2",
+    "prettier": "^2.8.3",
+    "webpack": "^5.75.0",
+    "webpack-cli": "^4.10.0",
     "webpack-merge": "^5.8.0"
   },
   "scripts": {
@@ -40,7 +39,7 @@
     "assert": "^2.0.0",
     "browserify-zlib": "^0.2.0",
     "buffer": "^6.0.3",
-    "css-loader": "^6.7.1",
+    "css-loader": "^6.7.3",
     "d3-dispatch": "^1.0.6",
     "d3-drag": "^1.2.5",
     "d3-polygon": "^1.0.6",
@@ -50,23 +49,23 @@
     "eslint-plugin-simple-import-sort": "^8.0.0",
     "fast-json-patch": "^3.1.1",
     "https-browserify": "^1.0.0",
-    "jquery": "^3.6.0",
     "json-stable-stringify": "^1.0.1",
     "md5": "^2.3.0",
+    "jquery": "^3.6.3",
     "ml-savitzky-golay-generalized": "^4.0.1",
     "rc-tree-select": "^1.12.13",
-    "react": "^16.14.0",
-    "react-devtools": "^4.24.4",
-    "react-dom": "^16.14.0",
+    "react": "^18.2.0",
+    "react-devtools": "^4.27.1",
+    "react-dom": "^18.2.0",
     "react-grid-layout": "0.16.6",
-    "react-modal": "^3.14.4",
-    "react-resize-detector": "^4.2.3",
+    "react-modal": "^3.16.1",
+    "react-resize-detector": "^7.1.2",
     "stream-browserify": "^3.0.0",
     "stream-http": "^3.2.0",
     "style-loader": "^3.3.1",
     "three": "^0.105.2",
     "url": "^0.11.0",
-    "util": "^0.12.4",
+    "util": "^0.12.5",
     "whatwg-fetch": "^3.6.2"
   }
 }
