@@ -10,7 +10,11 @@ describe('Test Setup', () => {
   });
 
   it('manual server reconnect', () => {
-    cy.visit('/')
+    cy.visit('/').wait(1000)
+    cy.contains('online').click()
+    cy.contains('offline').click()
+    cy.contains('online').click()
+    cy.contains('offline').click()
     cy.contains('online').click()
     cy.contains('offline').click()
     cy.contains('online')
@@ -25,7 +29,7 @@ describe('Test Setup', () => {
   it('env selection works', () => {
     cy.visit('/')
     cy.get('.rc-tree-select [title="main"]').should('exist')
-    cy.get('.rc-tree-select').contains('main').trigger('mouseover')
+    cy.get('.rc-tree-select').contains('main').trigger('mouseover').wait(100);
     cy.get('.rc-tree-select .rc-tree-select-selection__choice__remove').click({force: true})
     cy.get('.rc-tree-select [title="main"]').should('not.exist')
     cy.get('.rc-tree-select').click()
