@@ -20,21 +20,24 @@ describe('Properties Pane', () => {
 
       cy.get('input[value="initial"]').first().clear().type("changed{enter}")
       cy.get('.layout .react-grid-item .content-text').first().contains("Updated: Text input => changed")
+      cy.get('input[value="changed_updated"]')
 
       cy.get('input[value="12"]').first().clear().type("42{enter}")
       cy.get('.layout .react-grid-item .content-text').first().contains("Updated: Number input => 42")
+      cy.get('input[value="420"]')
 
       cy.get('button').contains("Start").click()
       cy.get('.layout .react-grid-item .content-text').first().contains("Updated: Button => clicked")
 
-      cy.get('input[type="checkbox"]').first().click()
+      cy.get('input[type="checkbox"]').first().should("be.checked").click()
       cy.get('.layout .react-grid-item .content-text').first().contains("Updated: Checkbox => False")
+      cy.get('input[type="checkbox"]').first().should("not.be.checked")
 
-      cy.get('select').first().select('Red').select('Blue')
+      cy.get('select').first().should("have.value", "1").select('Red')
       cy.get('.layout .react-grid-item .content-text').first().contains("Updated: Select => 0")
-
-      cy.get('select').first().select('Blue')
+      cy.get('select').first().should("have.value", "0").select('Blue')
       cy.get('.layout .react-grid-item .content-text').first().contains("Updated: Select => 2")
+      cy.get('select').first().should("have.value", "2")
 
   })
 
